# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "NeedsYouKind": {
        "description": "Which gate the open work belongs to, so a row can name the ask.",
        "enum": [
          "plan",
          "approval",
          "mixed"
        ],
        "title": "NeedsYouKind",
        "type": "string"
      },
...
      "PlanProgress": {
        "description": "The one badge a trip row carries. `none` renders as no badge at all.",
        "enum": [
          "none",
          "preparing",
          "planned",
          "booking"
        ],
        "title": "PlanProgress",
        "type": "string"
      },
...
      "ProblemOut": {
        "description": "The error body, declared so it reaches the generated contract.\n\nRendered by an exception handler rather than returned from a route, so\nFastAPI cannot infer it; app/api/router.py attaches it to every operation\nas the default response. problem_response() below builds from this model,\nwhich is what keeps the declaration and the wire body the same shape.",
        "properties": {
          "code": {
            "description": "Stable machine code the client maps to a failure state.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code",
          "type"
        ],
        "title": "ProblemOut",
        "type": "object"
      },
...
      "SourceKind": {
        "enum": [
          "google_calendar",
          "gmail",
          "apple_calendar",
          "manual_import"
        ],
        "title": "SourceKind",
        "type": "string"
      },
...
      "TimelineOut": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/TimelineEntryOut"
            },
            "title": "Entries",
            "type": "array"
          }
        },
        "required": [
          "entries"
        ],
        "title": "TimelineOut",
        "type": "object"
      },
      "TodayViewOut": {
        "properties": {
          "day_label": {
            "title": "Day Label",
            "type": "string"
          },
          "next_up": {
            "items": {
              "$ref": "#/components/schemas/PlanItemOut"
            },
            "title": "Next Up",
            "type": "array"
          },
          "state_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Detail"
          },
          "state_word": {
            "title": "State Word",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "trip_id": {
            "format": "uuid",
            "title": "Trip Id",
            "type": "string"
          },
          "window": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WellnessWindowOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "trip_id",
          "day_label",
          "state_word",
          "timezone",
          "next_up"
        ],
        "title": "TodayViewOut",
        "type": "object"
      },
      "TripConfirmIn": {
        "properties": {
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "updated_at"
        ],
        "title": "TripConfirmIn",
        "type": "object"
      },
      "TripCreateIn": {
        "properties": {
          "destination_name": {
            "title": "Destination Name",
            "type": "string"
          },
          "ends_on": {
            "format": "date",
            "title": "Ends On",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "lodging_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lodging Name"
          },
          "starts_on": {
            "format": "date",
            "title": "Starts On",
            "type": "string"
          }
        },
        "required": [
          "destination_name",
          "starts_on",
          "ends_on"
        ],
        "title": "TripCreateIn",
        "type": "object"
      },
      "TripEvidenceOut": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/SourceKind"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "source",
          "summary",
          "kind"
        ],
        "title": "TripEvidenceOut",
        "type": "object"
      },
      "TripListOut": {
        "properties": {
          "trips": {
            "items": {
              "$ref": "#/components/schemas/TripOut"
            },
            "title": "Trips",
            "type": "array"
          }
        },
        "required": [
          "trips"
        ],
        "title": "TripListOut",
        "type": "object"
      },
      "TripOrigin": {
        "enum": [
          "calendar_detection",
          "manual",
          "import"
        ],
        "title": "TripOrigin",
        "type": "string"
      },
      "TripOut": {
        "properties": {
          "destination_name": {
            "title": "Destination Name",
            "type": "string"
          },
          "detection_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detection Confidence"
          },
          "ends_on": {
            "format": "date",
            "title": "Ends On",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/TripEvidenceOut"
            },
            "title": "Evidence",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "needs_you_count": {
            "title": "Needs You Count",
            "type": "integer"
          },
          "needs_you_kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NeedsYouKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "origin": {
            "$ref": "#/components/schemas/TripOrigin"
          },
          "plan_progress": {
            "$ref": "#/components/schemas/PlanProgress"
          },
          "starts_on": {
            "format": "date",
            "title": "Starts On",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TripState"
          },
          "state_line": {
            "title": "State Line",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "origin",
          "destination_name",
          "timezone",
          "starts_on",
          "ends_on",
          "evidence",
          "state_line",
          "plan_progress",
          "needs_you_count",
          "updated_at"
        ],
        "title": "TripOut",
        "type": "object"
      },
      "TripState": {
        "enum": [
          "detected",
          "confirmed",
          "upcoming",
          "preparing",
          "active",
          "completed",
          "archived",
          "dismissed"
        ],
        "title": "TripState",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Generated from the FastAPI routers by backend/scripts/dump_openapi.py. Do not edit: run the script. This describes what the server actually serves and is what the frontend's types are generated from. The hand-written docs/openapi.yaml is the design artifact and may declare endpoints ahead of the code; backend/scripts/check_openapi_drift.py holds the two together. See docs/adr/004.",
    "title": "TravelWell API (served)",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/trips": {
      "get": {
        "operationId": "list_trips_api_v1_trips_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TripState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "List Trips",
        "tags": [
          "trips"
        ]
      },
      "post": {
        "operationId": "create_trip_api_v1_trips_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Create Trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{trip_id}": {
      "get": {
        "operationId": "get_trip_api_v1_trips__trip_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Get Trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{trip_id}/confirm": {
      "post": {
        "operationId": "confirm_trip_api_v1_trips__trip_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripConfirmIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Confirm Trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{trip_id}/dismiss": {
      "post": {
        "description": "\"Not a trip\". Detection is noisy, so the gate has to open both ways.",
        "operationId": "dismiss_trip_api_v1_trips__trip_id__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripConfirmIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Dismiss Trip",
        "tags": [
          "trips"
        ]
      }
    },
...
    "/trips/{trip_id}/timeline": {
      "get": {
        "operationId": "get_trip_timeline_api_v1_trips__trip_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Get Trip Timeline",
        "tags": [
          "trips"
        ]
      }
    },
    "/trips/{trip_id}/today": {
      "get": {
        "operationId": "get_trip_today_api_v1_trips__trip_id__today_get",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayViewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Problem details (RFC 9457)"
          }
        },
        "summary": "Get Trip Today",
        "tags": [
          "trips"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}